refactor signer related input parsers into a submodule
